tones sound a bit better now
I am almost positive that the reason we have to press a button multiple times is because when we read the button in it loops and loops resetting the pointer each time rather than setting it once and waiting. I tried to add some logic dealing with ring buffer length to fix it but it's not quite there yet.